various performance improvements and tidying up, moved vis and theta to main package modules
