+++ radd/fit.py
@@ -5,48 +5,13 @@ from copy import deepcopy
 import numpy as np
 import pandas as pd
 from numpy import array
-from radd.tools import messages, theta
+from radd import theta
+from radd.tools import messages
 from radd.tools.analyze import assess_fit
 from radd.models import Simulator
 from lmfit import minimize, fit_report
 from scipy.optimize import basinhopping
 from numpy.random import uniform
-from radd.tools.vis import PBinJ
-
-class BasinFeedback(object):
-    """ tracks BasinHopping progress with tqdm bars
-    Arguments:
-        nsuccess:
-            Stop the run if the global minimum candidate remains the
-            same for this number of iterations.
-    """
-    def __init__(self, ninits, fmin0=None):
-        self.inits_bar = PBinJ(bartype='colorbar', n=ninits+1, color='blue', title='global fmin')
-        self.success_bar = PBinJ(bartype='infobar', title='current fmin')
-        self.basins = []
-        if fmin0 is not None:
-            self.inits_bar.update(new_progress=fmin0)
-            self.basins.append(fmin0)
-
-    def callback(self, x, fmin, accept):
-        """ A callback function which will be called for all minima found
-        Arguments:
-            x (array):
-                parameter values
-            fmin (float):
-                function value of the trial minimum, and
-            accept (bool):
-                whether or not that minimum was accepted
-        """
-        if fmin <= np.min(self.basins):
-            self.inits_bar.update(new_progress=fmin)
-        if accept:
-            self.basins.append(fmin)
-            self.success_bar.update(new_progress=fmin)
-
-    def clear(self):
-        self.inits_bar.clear()
-        self.success_bar.clear()
 
 class BasinBounds(object):
     """ sets conditions for step acceptance during
@@ -108,45 +73,7 @@ class Optimizer(object):
         self.pnames = ['a', 'tr', 'v', 'ssv', 'z', 'xb', 'si', 'sso']
         self.constants = deepcopy(['a', 'tr', 'v', 'xb'])
 
-    def sample_inits(self, pkeys=None):
-        """ test a large sample of random parameter values
-        and submit <nkeep> to hop_around() global optimization
-        """
-        bp = self.basinparams
-        nkeep = bp['ninits']
-        nsamples = bp['nsamples']
-        keep_method = bp['init_sample_method']
-        if pkeys is None:
-            pkeys = np.sort(list(self.inits))
-        rinits = theta.random_inits(pkeys, ninits=nsamples, kind=self.kind)
-        all_inits = [{pk: rinits[pk][i] for pk in pkeys} for i in range(nsamples)]
-        fmin_all = [self.simulator.cost_fx(inits_i, sse=True) for inits_i in all_inits]
-        # rank inits by costfx error low-to-high
-        fmin_series = pd.Series(fmin_all)
-        rankorder = fmin_series.sort_values()
-        # eliminate extremely bad parameter sets
-        rankorder = rankorder[rankorder<=5.0]
-        if keep_method=='random':
-            # return nkeep from randomly sampled inits
-            inits = all_inits[:nkeep]
-            inits_err = fmin_all[:nkeep]
-        elif keep_method=='best':
-            # return nkeep from inits with lowest err
-            inits = [all_inits[i] for i in rankorder.index[:nkeep]]
-            inits_err = rankorder.values[:nkeep]
-        elif keep_method=='lmh':
-            # split index for low, med, and high err inits
-            # if nkeep is odd, will sample more low than high
-            if nkeep<3: nkeep=3
-            ix = rankorder.index.values
-            nl, nm, nh = [arr.size for arr in np.array_split(np.arange(nkeep), 3)]
-            # extract indices roughly equal numbers of parameter sets with low, med, hi err
-            keep_ix = np.hstack([ix[:nl], np.array_split(ix,2)[0][-nm:], ix[-nh:]])
-            inits = [all_inits[i] for i in keep_ix]
-            inits_err = [fmin_series[i] for i in keep_ix]
-        return inits, inits_err
-
-    def hop_around(self, p, inits_list=None, progress=False, callback=None):
+    def hop_around(self, inits_list, pbars=None):
         """ initialize model with niter randomly generated parameter sets
         and perform global minimization using basinhopping algorithm
         ::Arguments::
@@ -155,27 +82,21 @@ class Optimizer(object):
         ::Returns::
             parameter set with the best fit
         """
-        bp = self.basinparams
-        if inits_list is None:
-            # sample random inits and select best of
-            inits_list, fmins = self.sample_inits(pkeys=np.sort(list(p)))
-            f0 = np.min(fmins)
-        else:
-            f0=np.min([self.simulator.cost_fx(inits, sse=True) for inits in inits_list])
-        if progress and not bp['disp']:
-            pbars = BasinFeedback(ninits=len(inits_list), fmin0=f0)
+        callback = None
+        progress = False
+        if pbars is not None:
+            progress = True
         xpopt, xfmin = [], []
         for i, inits in enumerate(inits_list):
-            if progress and not bp['disp']:
-                pbars.inits_bar.update(i+1)
+            if progress:
+                pbars.update(name='glb_basin', i=i)
                 callback = pbars.callback
             popt, fmin = self.run_basinhopping(p=inits, is_flat=True, callback=callback)
             xpopt.append(popt)
             xfmin.append(fmin)
-        if progress:
-            pbars.clear()
-        # report global minimum
-        print("GLOBAL MIN = {:.9f}".format(np.min(xfmin)))
+        if self.fitparams['disp']:
+            # report global minimum
+            print("Finished Hopping Around:\nGLOBAL MIN = {:.9f}".format(np.min(xfmin)))
         # return parameters at the global basin
         return xpopt[np.argmin(xfmin)]
