+++ radd/theta.py
@@ -8,7 +8,7 @@ from numpy import array
 from scipy.stats.distributions import norm, gamma, uniform
 from lmfit import Parameters
 
-def random_inits(pkeys, ninits=1, kind='dpm', mu=None, sigma=None):
+def random_inits(pkeys, ninits=1, kind='dpm', mu=None, sigma=None, as_list_of_dicts=False):
     """ random parameter values for initiating model across range of
     parameter values.
     ::Arguments::
@@ -28,7 +28,39 @@ def random_inits(pkeys, ninits=1, kind='dpm', mu=None, sigma=None):
     if 'vd' in params.keys():
         # vi_perc ~ U(.05, .95) --> vi = vi_perc*vd
         params['vi'] = params['vi']*params['vd']
+    if as_list_of_dicts:
+        params = [{pk: params[pk][i] for pk in pkeys} for i in range(ninits)]
     return params
+
+def filter_param_sets(param_sets, param_yhats, fitparams, nkeep=5, keep_method='best'):
+    # make copies of params, yhats, and fitparams
+    p_sets, p_yhats, fp = [deepcopy(orig) for orig in [param_sets, param_yhats, fitparams]]
+    y = fp['y'].flatten();  wts = fp['wts'].flatten()
+    p_fmins = [np.sum((wts * (yhat - y))**2) for yhat in p_yhats]
+    # rank inits by costfx error low-to-high
+    fmin_series = pd.Series(p_fmins)
+    rankorder = fmin_series.sort_values()
+    # eliminate extremely bad parameter sets
+    rankorder = rankorder[rankorder<=5.0]
+    if keep_method=='random':
+        # return nkeep from randomly sampled inits
+        inits = p_sets[:nkeep]
+        inits_err = p_fmins[:nkeep]
+    elif keep_method=='best':
+        # return nkeep from inits with lowest err
+        inits = [p_sets[i] for i in rankorder.index[:nkeep]]
+        inits_err = rankorder.values[:nkeep]
+    elif keep_method=='lmh':
+        # split index for low, med, and high err inits
+        # if nkeep is odd, will sample more low than high
+        if nkeep<3: nkeep=3
+        ix = rankorder.index.values
+        nl, nm, nh = [arr.size for arr in np.array_split(np.arange(nkeep), 3)]
+        # extract indices roughly equal numbers of parameter sets with low, med, hi err
+        keep_ix = np.hstack([ix[:nl], np.array_split(ix,2)[0][-nm:], ix[-nh:]])
+        inits = [p_sets[i] for i in keep_ix]
+        inits_err = [fmin_series[i] for i in keep_ix]
+    return inits, np.min(inits_err)
 
 def loadParameters(inits=None, is_flat=False, kind=None, pc_map={}):
     """ Generates and returns an lmfit Parameters object with
